erlang.inc: integrate build fix from erlang_R16B.bb
The erlang_R16B03-1 build was failing due to a build issue that was
fixed with erlang-fix-build-issue-in-Yocto.patch, already present in
the erlang_R15B build. Promote it to the shared .inc file. This means
the patch also applies to erlang-native, but unless you're building
under Solaris and your libdlpi is installed in a nonstandard location,
you're probably not going to notice it.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>